moved settings files + stripJsonComments
