+++ extension/package.json
@@ -2,7 +2,7 @@
     "name": "nab-al-tools",
     "displayName": "NAB AL Tools",
     "description": "Development and translation management tool for AL language and Microsoft Dynamics 365 Business Central extensions.",
-    "version": "1.2.0-preview.3",
+    "version": "1.2.0-preview.7",
     "publisher": "nabsolutions",
     "repository": {
         "type": "git",
@@ -32,7 +32,7 @@
             "url": "https://www.nabsolutions.se"
         }
     ],
-    "icon": "images/icon.png",
+    "icon": "icon.png",
     "bugs": {
         "url": "https://github.com/jwikman/nab-al-tools/issues"
     },
@@ -590,6 +590,7 @@
         "node-powershell": "^3.3.1",
         "original-fs": "^1.1.0",
         "semver": "^7.3.5",
+        "strip-json-comments": "^3.1.1",
         "time-stamp": "^1.1.0",
         "txml": "^4.0.1",
         "xmldom": "^0.5.0"
@@ -606,6 +607,7 @@
         "@types/node": "^12.20.12",
         "@types/node-powershell": "^3.1.1",
         "@types/semver": "^7.3.5",
+        "@types/strip-json-comments": "^3.0.0",
         "@types/vscode": "^1.50.0",
         "@types/xmldom": "^0.1.30",
         "@typescript-eslint/eslint-plugin": "^4.22.1",
